### 2017-10-09 #428 @whedon
[PRE REVIEW]: PySwarms: a research toolkit for Particle Swarm Optimization in Python
**Submitting author:** @ljvmiranda921 (<a href="http://orcid.org/0000-0002-7872-6464">Lester James Miranda</a>)
**Repository:** <a href="https://github.com/ljvmiranda921/pyswarms" target ="_blank">https://github.com/ljvmiranda921/pyswarms</a>
**Version:** v0.1.7
