### 2017-10-14 #428 @ljvmiranda921
Hi @kyleniemeyer , thank you for taking the time to oversee the review of this submission.

I checked the [list of reviewers for JOSS](https://github.com/openjournals/joss/blob/master/docs/reviewers.csv), then looked for keywords such as "optimization" and "python". I think that @stsievert and @oesteban may be suitable to review this if they have the time.

Thank you so much!
