### 2017-10-17 #428 @stsievert
Ditto. Give me another two weeks.

@ljvmiranda921 [hyperopt] and the [accompanying paper] might be of interest for the hyper-parameter optimization. Also, [dask-searchcv] might be of interest.

[hyperopt]:https://github.com/hyperopt/hyperopt
[accompanying paper]:https://papers.nips.cc/paper/4443-algorithms-for-hyper-parameter-optimization.pdf
[dask-searchcv]:https://github.com/dask/dask-searchcv
