### 2017-10-17 #428 @ljvmiranda921
Sure, thanks for accepting to review @oesteban , @stsievert . I'd be waiting for your review at the time you've mentioned.

Okay, in the meantime I will check on the hyperparameter optimization libraries stsievert mentioned. It may be useful to add an adapter to one of those libraries in pyswarms.

Thanks a lot!

cc: @kyleniemeyer
