### 2017-10-18 #428 @kyleniemeyer
Hi @stsievert and @oesteban, thanks for agreeing to review! (Hopefully that's what you meant with those messages 😄). I'll get the actual review issue started, with @stsievert set as the "main" reviewer since he expects to get to it a bit sooner—but both of you can add comments. (JOSS currently only supports one official reviewer, though we hope to fix that soon.)

@ljvmiranda921 I appreciate your effort finding reviewers, but I will shepherd the review... less appropriate for the submitting author to do so 😆
