### 2017-10-18 #433 @whedon
[REVIEW]: PySwarms: a research toolkit for Particle Swarm Optimization in Python
**Submitting author:** @ljvmiranda921 (<a href="http://orcid.org/0000-0002-7872-6464">Lester James Miranda</a>)
**Repository:** <a href="https://github.com/ljvmiranda921/pyswarms" target ="_blank">https://github.com/ljvmiranda921/pyswarms</a>
**Version:** v0.1.7

- [x] **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license? [Yes, MIT]
- [x] **Version:** Does the release version given match the GitHub release (v0.1.7)? [Yes, a prerelease]

- [x] **Installation:** Does installation proceed as outlined in the documentation? [Yes, worked for me]
- [x] **Functionality:** Have the functional claims of the software been confirmed? [Yes, docs make it clear that classes have parameters]
- [x] **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.) [I see no performance claims]

- [x] **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution. [setup.py, requirements_dev.txt]
- [x] **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems). [Yes, in README and examples/]
- [x] **Automated tests:** Are there automated tests or manual steps described so that the function of the software can be verified? [Yes, TravisCI]
- [x] **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support [Yes, and linked on docs]
