### 2017-10-19 #428 @stsievert
> Hopefully that's what you meant with those messages

Yup, that's what I meant (but can't get to it right away).

> (JOSS currently only supports one official reviewer, though we hope to fix that soon.)

I also ran into this in #382. Hopefully it's fixed soon!
