### 2017-10-28 #433 @stsievert
I've looked this over briefly today, and here's some comments:

* I didn't see a statement of need that clearly addressed the audience. Is the audience swarm intelligence researchers? Other uses of neural nets that would some other optimization algorithm?
* Particle Swarm Optimization is mentioned many times, but I've never heard of it. Is be used in place of sklearn's GridSearch (i.e., is it useful in hyper parameter tuning for machine learning)? I'd like to see a brief description of what (if any?) guarantees PSOs have.
* What's the difference in the packages for global min and local min? If minimizing a convex function these two packages will produce the same result. Are you most concerned about non-convex functions?
