### 2017-10-29 #433 @ljvmiranda921
Hi @stsievert ,

Thank you so much for your comments.

An aside, PSO is a class of search algorithms that can be used to optimize functions even if they're differentiable or not. This library provides implementations of these algorithms that should solve any optimization problem. Of course, the performance will depend on the PSO parameters (you can control if the particles are independent from one another, follows the herd, etc.), and to the type of PSO used in a certain problem.

As for your comments:

- The package is intended for swarm int. researchers & students who would like a high-level interface of implementing PSO in their problems. It doesn't have to be a neural nets problem, in this [use-case](https://pyswarms.readthedocs.io/en/latest/examples/usecases.html), it demonstrates how the parameters of a neural-net can be optimized through this package. It also doesn't need to be a ML problem, as long as an objective function is defined, then it can be used as input in pyswarms. This can range from performing inverse-kinematics in robotics, or job-shop scheduling (It just so happened that most of my use-cases are related to ML).
- Although you can use swarm intelligence algorithms for finding the hyperparameters in a machine learning classifier [1] [2], similar to GridSearch/RandomSearch in sklearn, the GridSearch and RandomSearch in this package is only for finding the "hyperparameters" for the PSO algorithm (a.k.a. the parameters that controls its behaviour).  In a vanilla PSO, you have parameters that controls how much the particles are dependent from one another, or how much of their neighbors they're going to follow.
- I believe you're referring to `GlobalBestPSO` and `LocalBestPSO`? This doesn't necessarily refer to global minima or local minima. These two define the "topology" of the swarm. In `GlobalBestPSO`, all particles follow a *leader*, who has the best cost in the current iteration [3]. In `LocalBestPSO`, the particles are divided into neighbors/clusters, and only looks at the *sub-leader* in their respective neighborhoods [4]. I'm using the words `GlobalBest` and `LocalBest` because it's the one used in literature [3] [4].

Thank you so much, I hope it clears things up! 😄

---
[1] Dario Floreano, Peter Durr, and Claudio Mattiussi. “Neuroevolution: from architectures to
learning”. In: Evolutionary Intelligence (2008).
[2] Kenneth Stanley, David B. D’Ambrosio, and Jason Gauci. “A hypercube-based encoding
for evolving large-scale neural networks”. In: Artificial Life (2009).
[3] J. Kennedy and R.C. Eberhart, “Particle Swarm Optimization,” Proceedings of the IEEE International Joint Conference on Neural Networks, 1995, pp. 1942-1948.
[4] Y. Shi and R.C. Eberhart, “A modified particle swarm optimizer,” Proceedings of the IEEE International Conference on Evolutionary Computation, 1998, pp. 69-73.
