### 2017-10-29 #433 @stsievert
> The package is intended for swarm int. researchers & students who would like a high-level interface of implementing PSO in their problems.

Thanks – that clears this up. I don't think the above is expressed clearly in your docs. I would like to see the same sentence, maybe with "high-level interface" expanded.

> Thank you so much, I hope it clears things up!

It does! Thank you.

[hyperband]:http://people.eecs.berkeley.edu/~kjamieson/hyperband.html
