### 2017-10-29 e56e4cd
Fix PEP 8 Violations (#54)
Three violations remain:

pyswarms/discrete/binary.py:17:80: E501 line too long (117 > 79 characters)
pyswarms/single/global_best.py:23:80: E501 line too long (117 > 79 characters)
pyswarms/single/local_best.py:23:80: E501 line too long (117 > 79 characters)

are overlong `.. math::` lines within docstrings.

Author: jazcap53
Email: andrew.jarcho@gmail.com
