+++ pyswarms/single/global_best.py
@@ -4,8 +4,8 @@ r"""
 A Global-best Particle Swarm Optimization (gbest PSO) algorithm.
 
 It takes a set of candidate solutions, and tries to find the best
-solution using a position-velocity update method. Uses a 
-star-topology where each particle is attracted to the best 
+solution using a position-velocity update method. Uses a
+star-topology where each particle is attracted to the best
 performing particle.
 
 The position update can be defined as:
@@ -27,7 +27,7 @@ respectively. They control the particle's behavior in choosing how to
 react given two choices: (1) to follow its *personal best* or (2) follow
 the swarm's *global best* position. Overall, this dictates if the swarm
 is explorative or exploitative in nature. In addition, a parameter
-:math:`w` controls the inertia of the swarm's movement. 
+:math:`w` controls the inertia of the swarm's movement.
 
 An example usage is as follows:
 
@@ -40,7 +40,8 @@ An example usage is as follows:
     options = {'c1': 0.5, 'c2': 0.3, 'w':0.9}
 
     # Call instance of GlobalBestPSO
-    optimizer = ps.single.GlobalBestPSO(n_particles=10, dimensions=2, options=options)
+    optimizer = ps.single.GlobalBestPSO(n_particles=10, dimensions=2,
+                                        options=options)
 
     # Perform optimization
     stats = optimizer.optimize(fx.sphere_func, iters=100)
@@ -67,11 +68,12 @@ from past.builtins import xrange
 from ..base import SwarmBase
 from ..utils.console_utils import cli_print, end_report
 
+
 class GlobalBestPSO(SwarmBase):
 
-    def __init__(self, n_particles, dimensions, options, 
-        bounds=None, velocity_clamp=None):
-        """Initializes the swarm. 
+    def __init__(self, n_particles, dimensions, options,
+                 bounds=None, velocity_clamp=None):
+        """Initializes the swarm.
 
         Attributes
         ----------
@@ -80,7 +82,7 @@ class GlobalBestPSO(SwarmBase):
         dimensions : int
             number of dimensions in the space.
         options : dict with keys :code:`{'c1', 'c2', 'w'}`
-            a dictionary containing the parameters for the specific 
+            a dictionary containing the parameters for the specific
             optimization technique
                 * c1 : float
                     cognitive parameter
@@ -94,10 +96,11 @@ class GlobalBestPSO(SwarmBase):
             be of shape :code:`(dimensions,)`.
         velocity_clamp : tuple (default is :code:`None`)
             a tuple of size 2 where the first entry is the minimum velocity
-            and the second entry is the maximum velocity. It 
-            sets the limits for velocity clamping. 
+            and the second entry is the maximum velocity. It
+            sets the limits for velocity clamping.
         """
-        super(GlobalBestPSO, self).__init__(n_particles, dimensions, options, bounds, velocity_clamp)
+        super(GlobalBestPSO, self).__init__(n_particles, dimensions, options,
+                                            bounds, velocity_clamp)
 
         # Initialize logger
         self.logger = logging.getLogger(__name__)
@@ -116,8 +119,8 @@ class GlobalBestPSO(SwarmBase):
         ----------
         objective_func : function
             objective function to be evaluated
-        iters : int 
-            number of iterations 
+        iters : int
+            number of iterations
         print_step : int (default is 1)
             amount of steps for printing into console.
         verbose : int  (default is 1)
@@ -138,8 +141,9 @@ class GlobalBestPSO(SwarmBase):
             m = (current_cost < pbest_cost)
             pbest_cost = np.where(~m, pbest_cost, current_cost)
             # Create 2-D mask to update positions
-            _m = np.repeat(m[:,np.newaxis], self.dimensions, axis=1)
[…]
