+++ pyswarms/base/base_discrete.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 
 r"""
-Base class for single-objective discrete Particle Swarm Optimization 
+Base class for single-objective discrete Particle Swarm Optimization
 implementations.
 
-All methods here are abstract and raises a :code:`NotImplementedError` 
+All methods here are abstract and raises a :code:`NotImplementedError`
 when not used. When defining your own swarm implementation,
 create another class,
 
@@ -35,11 +35,12 @@ import numpy as np
 import logging.config
 from collections import namedtuple
 
+
 class DiscreteSwarmBase(object):
 
     def assertions(self):
         """Assertion method to check various inputs.
-        
+
         Raises
         ------
         TypeError
@@ -58,16 +59,18 @@ class DiscreteSwarmBase(object):
             if not isinstance(self.velocity_clamp, tuple):
                 raise TypeError('Parameter `velocity_clamp` must be a tuple')
             if not len(self.velocity_clamp) == 2:
-                raise IndexError('Parameter `velocity_clamp` must be of size 2')
+                raise IndexError('Parameter `velocity_clamp` must be of '
+                                 'size 2')
             if not self.velocity_clamp[0] < self.velocity_clamp[1]:
-                raise ValueError('Make sure that velocity_clamp is in the form (v_min, v_max)')
+                raise ValueError('Make sure that velocity_clamp is in the '
+                                 'form (v_min, v_max)')
 
         # Required keys in options argument
-        if not all (key in self.options for key in ('c1', 'c2', 'w')):
+        if not all(key in self.options for key in ('c1', 'c2', 'w')):
             raise KeyError('Missing either c1, c2, or w in options')
 
-    def setup_logging(self, default_path='./config/logging.yaml', 
-            default_level=logging.INFO, env_key='LOG_CFG'):
+    def setup_logging(self, default_path='./config/logging.yaml',
+                      default_level=logging.INFO, env_key='LOG_CFG'):
         """Setup logging configuration
 
         Parameters
@@ -90,10 +93,9 @@ class DiscreteSwarmBase(object):
         else:
             logging.basicConfig(level=default_level)
 
-
-    def __init__(self, n_particles, dimensions, binary, options, 
-        velocity_clamp=None):
-        """Initializes the swarm. 
+    def __init__(self, n_particles, dimensions, binary, options,
+                 velocity_clamp=None):
+        """Initializes the swarm.
 
         Creates a :code:`numpy.ndarray` of positions depending on the
         number of particles needed and the number of dimensions.
@@ -112,10 +114,10 @@ class DiscreteSwarmBase(object):
             random integers from 0 to :code:`dimensions` are generated.
         velocity_clamp : tuple (default is :code:`None`)
             a tuple of size 2 where the first entry is the minimum velocity
-            and the second entry is the maximum velocity. It 
-            sets the limits for velocity clamping. 
+            and the second entry is the maximum velocity. It
+            sets the limits for velocity clamping.
         options: dict
-            a dictionary containing the parameters for a specific 
+            a dictionary containing the parameters for a specific
             optimization technique
         """
         self.setup_logging()
@@ -127,8 +129,10 @@ class DiscreteSwarmBase(object):
         self.swarm_size = (n_particles, dimensions)
         self.options = options
         # Initialize named tuple for populating the history list
-        self.ToHistory = namedtuple('ToHistory', ['best_cost', 'mean_pbest_cost',
-                            'mean_neighbor_cost', 'position', 'velocity'])
+        self.ToHistory = namedtuple('ToHistory',
+                                    ['best_cost', 'mean_pbest_cost',
+                                     'mean_neighbor_cost', 'position',
+                                     'velocity'])
         # Invoke assertions
[…]
