+++ pyswarms/single/local_best.py
@@ -5,7 +5,7 @@ A Local-best Particle Swarm Optimization (lbest PSO) algorithm.
 
 Similar to global-best PSO, it takes a set of candidate solutions,
 and finds the best solution using a position-velocity update method.
-However, it uses a ring topology, thus making the particles 
+However, it uses a ring topology, thus making the particles
 attracted to its corresponding neighborhood.
 
 The position update can be defined as:
@@ -45,20 +45,22 @@ An example usage is as follows:
 
     # Call instance of LBestPSO with a neighbour-size of 3 determined by
     # the L2 (p=2) distance.
-    optimizer = ps.single.LocalBestPSO(n_particles=10, dimensions=2, options=options)
+    optimizer = ps.single.LocalBestPSO(n_particles=10, dimensions=2,
+                                       options=options)
 
     # Perform optimization
     stats = optimizer.optimize(fx.sphere_func, iters=100)
 
 This algorithm was adapted from one of the earlier works of
-J. Kennedy and R.C. Eberhart in Particle Swarm Optimization [IJCNN1995]_ [MHS1995]
+J. Kennedy and R.C. Eberhart in Particle Swarm Optimization
+[IJCNN1995]_ [MHS1995]
 
 .. [IJCNN1995] J. Kennedy and R.C. Eberhart, "Particle Swarm Optimization,"
     Proceedings of the IEEE International Joint Conference on Neural
     Networks, 1995, pp. 1942-1948.
 
 .. [MHS1995] J. Kennedy and R.C. Eberhart, "A New Optimizer using Particle
-    Swarm Theory,"  in Proceedings of the Sixth International 
+    Swarm Theory,"  in Proceedings of the Sixth International
     Symposium on Micromachine and Human Science, 1995, pp. 39–43.
 """
 
@@ -77,6 +79,7 @@ from past.builtins import xrange
 from ..base import SwarmBase
 from ..utils.console_utils import cli_print, end_report
 
+
 class LocalBestPSO(SwarmBase):
 
     def assertions(self):
@@ -93,14 +96,17 @@ class LocalBestPSO(SwarmBase):
         """
         super(LocalBestPSO, self).assertions()
 
-        if not all (key in self.options for key in ('k', 'p')):
+        if not all(key in self.options for key in ('k', 'p')):
             raise KeyError('Missing either k or p in options')
         if not 0 <= self.k <= self.n_particles:
-            raise ValueError('No. of neighbors must be between 0 and no. of particles.')
-        if self.p not in [1,2]:
-            raise ValueError('p-value should either be 1 (for L1/Minkowski) or 2 (for L2/Euclidean).')
+            raise ValueError('No. of neighbors must be between 0 and no. '
+                             'of particles.')
+        if self.p not in [1, 2]:
+            raise ValueError('p-value should either be 1 (for L1/Minkowski) '
+                             'or 2 (for L2/Euclidean).')
 
-    def __init__(self, n_particles, dimensions, options, bounds=None, velocity_clamp=None):
+    def __init__(self, n_particles, dimensions, options, bounds=None,
+                 velocity_clamp=None):
         """Initializes the swarm.
 
         Attributes
@@ -115,10 +121,10 @@ class LocalBestPSO(SwarmBase):
             be of shape :code:`(dimensions,)`.
         velocity_clamp : tuple (default is :code:`None`)
             a tuple of size 2 where the first entry is the minimum velocity
-            and the second entry is the maximum velocity. It 
-            sets the limits for velocity clamping. 
+            and the second entry is the maximum velocity. It
+            sets the limits for velocity clamping.
         options : dict with keys :code:`{'c1', 'c2', 'w', 'k', 'p'}`
-            a dictionary containing the parameters for the specific 
[…]
