+++ pyswarms/discrete/binary.py
@@ -58,7 +58,7 @@ from __future__ import print_function
 
 # Import modules
 import logging
-import numpy as np 
+import numpy as np
 from scipy.spatial import cKDTree
 from past.builtins import xrange
 
@@ -66,8 +66,9 @@ from past.builtins import xrange
 from ..base import DiscreteSwarmBase
 from ..utils.console_utils import cli_print, end_report
 
+
 class BinaryPSO(DiscreteSwarmBase):
-    
+
     def assertions(self):
         """Assertion method to check various inputs.
 
@@ -82,12 +83,14 @@ class BinaryPSO(DiscreteSwarmBase):
         """
         super(BinaryPSO, self).assertions()
 
-        if not all (key in self.options for key in ('k', 'p')):
+        if not all(key in self.options for key in ('k', 'p')):
             raise KeyError('Missing either k or p in options')
         if not 0 <= self.k <= self.n_particles:
-            raise ValueError('No. of neighbors must be between 0 and no. of particles.')
-        if self.p not in [1,2]:
-            raise ValueError('p-value should either be 1 (for L1/Minkowski) or 2 (for L2/Euclidean).')
+            raise ValueError('No. of neighbors must be between 0 and no. of'
+                             'particles.')
+        if self.p not in [1, 2]:
+            raise ValueError('p-value should either be 1 (for L1/Minkowski)'
+                             'or 2 (for L2/Euclidean).')
 
     def __init__(self, n_particles, dimensions, options, velocity_clamp=None):
         """Initializes the swarm.
@@ -100,10 +103,10 @@ class BinaryPSO(DiscreteSwarmBase):
             number of dimensions in the space.
         velocity_clamp : tuple (default is :code:`None`)
             a tuple of size 2 where the first entry is the minimum velocity
-            and the second entry is the maximum velocity. It 
-            sets the limits for velocity clamping. 
+            and the second entry is the maximum velocity. It
+            sets the limits for velocity clamping.
         options : dict with keys :code:`{'c1', 'c2', 'k', 'p'}`
-            a dictionary containing the parameters for the specific 
+            a dictionary containing the parameters for the specific
             optimization technique
                 * c1 : float
                     cognitive parameter
@@ -111,12 +114,12 @@ class BinaryPSO(DiscreteSwarmBase):
                     social parameter
                 * w : float
                     inertia parameter
-                * k : int 
+                * k : int
                     number of neighbors to be considered. Must be a
                     positive integer less than :code:`n_particles`
                 * p: int {1,2}
-                    the Minkowski p-norm to use. 1 is the 
-                    sum-of-absolute values (or L1 distance) while 2 is 
+                    the Minkowski p-norm to use. 1 is the
+                    sum-of-absolute values (or L1 distance) while 2 is
                     the Euclidean (or L2) distance.
         """
         # Initialize logger
@@ -126,7 +129,8 @@ class BinaryPSO(DiscreteSwarmBase):
         # Assign k-neighbors and p-value as attributes
         self.k, self.p = options['k'], options['p']
         # Initialize parent class
-        super(BinaryPSO, self).__init__(n_particles, dimensions, binary, options, velocity_clamp)
+        super(BinaryPSO, self).__init__(n_particles, dimensions, binary,
+                                        options, velocity_clamp)
         # Invoke assertions
         self.assertions()
         # Initialize the resettable attributes
@@ -166,19 +170,21 @@ class BinaryPSO(DiscreteSwarmBase):
             pbest_cost = np.where(~m, pbest_cost, current_cost)
             # Create a 2-D mask to update positions
             _m = np.repeat(m[:, np.newaxis], self.dimensions, axis=1)
[…]

[…]
