### 2017-10-30 fac00eb
Use SVG file instead of PNG
