### 2017-10-30 2942bf9
Add Gitter badge (#55)
This commit adds a gitter badge and removes "unnecessary" badges such as Health and pyup status
that I believe will just confuse the user.
