### 2017-10-30 ljvmiranda921/pyswarms#56 @ljvmiranda921
More explicit documentation
- [x] Clarify the statement of need, explain what a "high-level interface" is in README.rst, ReadTheDocs, and Paper.md
- [x] I think it would be better to explicitly acknowledge the v.0.1.7 committers aside from the README.rst, and their respective Issues and PRs in the Changelog. And if possible, in the Paper.md
