### 2017-10-30 ljvmiranda921/pyswarms#56 @ljvmiranda921
Reference: openjournals/joss-reviews#433
