### 2017-11-01 1af4006
Update History (#56)
This commit updates the format of HISTORY.rst. Here,
the actual Issue numbers and PRs are recorded for each
change. Proper credit also goes to the contributors for
each PR.

Author: ljvmiranda921
E-mail: ljvmiranda@gmail.com
