### 2017-11-01 129e038
Update paper.md (#56)
This commit updates paper.md in order to explicitly specify the Statement of Need of the software.

Author: ljvmiranda921
E-mail: ljvmiranda@gmail.com
