### 2017-11-01 b60247c
Update README.rst (#56)
Add Statement of Need

Author: ljvmiranda921
E-mail: ljvmiranda@gmail.com
