### 2017-11-01 #433 @ljvmiranda921
Hi @stsievert , I've made the changes you've requested both in Paper.md and the README.rst. Hope it's much better now. Thanks a lot!
