### 2017-11-01 ljvmiranda921/pyswarms#56 @stsievert
#57 adds this sentence:

> This means that instead of writing an optimizer from scratch, one just needs to supply the necessary PSO parameters, and let PySwarms do the rest of the procedure.

Is PySwarms composed of classes or functions? Here's a suggestion

> PySwarms allows both performing basic optimizations with PSO and interacting with PSO optimizations. Interaction is enabled because PySwarms provides object primitives required for PSO optimization. This makes PySwarms useful for researchers or students.

I'm sure this needs more edits.
