### 2017-11-01 #433 @stsievert
Thanks – this looks better! I've commented in https://github.com/ljvmiranda921/pyswarms/issues/56 for some more small edits.
