### 2017-11-01 ljvmiranda921/pyswarms#56 @ljvmiranda921
Hi @stsievert , I like your suggestion. Do you mean that I change

> It is intended for swarm intelligence researchers, practitioners, and students who would like a high-level interface of implementing PSO in their problems. This means that instead of writing an optimizer from scratch, one just needs to supply the necessary PSO parameters, and let PySwarms do the rest of the procedure.

Into this?

> PySwarms allows both performing basic optimizations with PSO and interacting with PSO optimizations. Interaction is enabled because PySwarms provides object primitives required for PSO optimization. This makes PySwarms useful for researchers or students.

Just a small edit for flow, how about something like [...?]:

>  It is intended for swarm intelligence researchers, practitioners, and students who would like a high-level interface of implementing PSO in their problems.  PySwarms both allows basic optimization with PSO and interaction with swarm optimizations. Interaction is enabled due to object primitives provided by the package for optimization. This makes PySwarms useful for researchers or students.

I just modified "PSO optimization", because it might read a little awkward if "PSO optimization" = "particle swarm optimization _optimization_"

Thank you so much for your suggestion and help 👍
