### 2017-11-01 ljvmiranda921/pyswarms#56 @stsievert
> and students who would like a high-level interface of implementing PSO

Small wording suggestion: maybe change "high-level interface" to "declarative interface"? Or maybe "high-level declarative interface"?

> a small edit for flow

Yup, looks good!

> "PSO optimization" ... "particle swarm optimization optimization"

Yeah, I wondering about that as I was typing.
