### 2017-11-01 ljvmiranda921/pyswarms#56 @ljvmiranda921
Awesome! Okay, so it becomes

> It is intended for swarm intelligence researchers, practitioners, and students who would like a high-level declarative interface of implementing PSO in their problems. PySwarms both allows basic optimization with PSO and interaction with swarm optimizations. Interaction is enabled due to object primitives provided by the package for optimization. This makes PySwarms useful for researchers or students.
