### 2017-11-01 ljvmiranda921/pyswarms#56 @stsievert
This looks good to me!
