### 2017-11-01 ljvmiranda921/pyswarms#56 @stsievert
Will README.rst also update the docs? I'd like to see an update to the docs too.
