### 2017-11-01 #433 @kyleniemeyer
@stsievert thanks for your comments so far! I noticed that a few of the checkboxes at the top are still empty, is your review still in progress?

(Also, please do check off that first one about the conflict of interest)
