### 2017-11-02 b2ec55f
Update Statement of Need (#56)
This commit updates the Statement of Need as discussed in Issue #56
in the Readme, Documentation, and Paper.

Author: ljvmiranda921
E-mail: ljvmiranda@gmail.com
