### 2017-11-02 402d5c3
Add Python 2.7 in required versions to pass
This updates the contribution guidelines, CONTRIBUTING.rst, to
include Python 2.7 to the required versions to pass for contributors.

Author: ljvmiranda921
E-mail: ljvmiranda@gmail.com
