+++ CONTRIBUTING.rst
@@ -102,6 +102,6 @@ Before you submit a pull request, check that it meets these guidelines:
 2. If the pull request adds functionality, the docs should be updated. Put
    your new functionality into a function with a docstring, and add the
    feature to the list in README.rst.
-3. The pull request should work for Python 3.4, 3.5, and above. Check
+3. The pull request should work for Python 2.7, 3.4, 3.5, and above. Check
    https://travis-ci.org/ljvmiranda921/pyswarms/pull_requests
    and make sure that the tests pass for all supported Python versions.
