### 2017-11-02 ljvmiranda921/pyswarms#56 @ljvmiranda921
Hi @stsievert , I didn't link README.rst with the docs so I edited them separately. Either way, the docs, README, and the Paper has been updated by commit fd00cc8582aeee370325c74b40b05ad313f3113a 😄
