### 2017-11-02 #433 @stsievert
@kyleniemeyer my review is still in progress. I still need to download and verify the claims in the documentation. Plus I'd like @oesteban's comment too.
