### 2017-11-02 ljvmiranda921/pyswarms#61 @stsievert
[DOC] Allow easier access to different doc pages
Going from two pages on the documentation requires going to an intermediate page. For example, going from [discrete.binary] to [single.global_best module] requires two clicks.

On the discrete.binary page, I see no link to single.global_best. That's illustrated by the figures below:

| discrete.binary | single.global_best |
| :--:|:---:|
| <img width="299" alt="screen shot 2017-11-02 at 2 43 31 pm" src="https://user-images.githubusercontent.com/1320475/32346750-50e57be6-bfdc-11e7-8295-8800a577bca5.png"> |<img width="299" alt="screen shot 2017-11-02 at 2 43 41 pm" src="https://user-images.githubusercontent.com/1320475/32346751-50f4e3a6-bfdc-11e7-8b96-7f99a2225148.png"> |

It would be useful to have links in-between sections.

This isn't blocking for https://github.com/openjournals/joss-reviews/issues/433 but it'd make docs for this package more accessible.

[discrete.binary]:https://pyswarms.readthedocs.io/en/latest/api/pyswarms.discrete.html#module-pyswarms.discrete.binary

The documentation would be easier to use if this was resolved.
