### 2017-11-02 ljvmiranda921/pyswarms#61 @stsievert
I think this could be resolved by calling "API Documentation" a section (not a subsection)
