### 2017-11-03 ljvmiranda921/pyswarms#61 @ljvmiranda921
Hi @stsievert , sorry I just saw this now. Thank you for the feedback and I’ll work on this right away.
