### 2017-11-06 af26420
Update API Docs navigation (#61)
This commit modifies the TOC tree and moves the API documentation
into its own Section (not a Subsection).Turns out that the submodules
header is unnecessary and can be omitted.

Reference: openjournals/joss-reviews#433

Author: ljvmiranda921
E-mail: ljvmiranda@gmail.com
