### 2017-11-06 e2975b8
Update API Docs navigation (#61)
This commit modifies the TOC tree and moves the API documentation
into its own Section (not a Subsection).Turns out that the submodules
header is unnecessary and can be omitted.

Author: ljvmiranda921
E-mail: ljvmiranda@gmail.com
