+++ docs/api/_pyswarms.optimizers.rst
@@ -1,0 +1,13 @@
+Optimizers
+==========
+
+The optimizers include the actual PSO implementations for various tasks.
+These include easy, off-the-shelf implementations include those that are
+already considered as standard in literature. This may include the
+classics such as global-best and local-best. Useful for quick-and-easy
+optimization problems.
+
+.. toctree::
+
+   pyswarms.single
+   pyswarms.discrete
