### 2017-11-08 ljvmiranda921/pyswarms#61 @ljvmiranda921
Hi @stsievert ,

I accomplished moving the **API Documentation** subsection in its own section, which I think is a really good suggestion, but the problem

> Going from two pages on the documentation requires going to an intermediate page. For example, going from discrete.binary to [single.global_best module] requires two clicks.

still persists. It turns out that one really has to click the specific package (`pyswarms.single` or `pyswarms.discrete`) to access the modules inside of it.

I'm thinking of removing the upper-level hierarchy (`pyswarms.single`, `pyswarms.discrete`), but I prefer PR #64 's set-up more because it divides the optimizers wrt the specific problem they're trying to solve (i.e. single-objective, discrete-optim). This also provides a good structure in case we expand to different problem types (e.g. `pyswarms.multi` or `pyswarms.constrained` for multi-objective and constrained).

I'm still looking for a JS workaround to automatically expose all the modules right away without having to click, but it might take more time.

### Move API Documentation as a new Section
![new toc](https://user-images.githubusercontent.com/12949683/32529712-62662410-c47d-11e7-8880-9ae0e5c3c669.PNG)

### Comparison
| discrete.binary | single.global_best |
| :--:|:---:|
| <img width="299" alt="pyswarms toc discrete" src="https://user-images.githubusercontent.com/12949683/32529659-0f2c4e3c-c47d-11e7-9da5-6bc952530f10.PNG"> |<img width="299" alt="pyswarms toc single" src="https://user-images.githubusercontent.com/12949683/32529662-14f32ec6-c47d-11e7-9ce1-88c3f26dc29b.PNG"> |
