### 2017-11-08 ljvmiranda921/pyswarms#61 @stsievert
A JS solution feels like a hack, and it looks like reduced the number of clicks required to navigate (ie, going from optimizers to utilizes looks easier).
