### 2017-11-08 ljvmiranda921/pyswarms#61 @ljvmiranda921
Hi @stsievert ,

> A JS solution feels like a hack

Well I feel the same way. I think it's part of Sphinx that the dropdown only appears depending on the page you are in.

Another way would be to remove the package hierarchy altogether (`single`, `discrete`), so as to expose the modules inside (`global_best`, `local_best`, `discrete`)&mdash; but I believe it's better to maintain the current structure, so the problem domains are much clearer (and scalable if new variants come in the future). What do you think?

> it looks like reduced the number of clicks required to navigate (ie, going from optimizers to utilizes looks easier).

Yes it does, your suggestion to move API Documentation up one section is actually good. Thanks! 👍
