### 2017-11-08 ljvmiranda921/pyswarms#61 @stsievert
> to remove the package hierarchy altogether (single, discrete),

I think the hierarchy should stay, it sounds like it adds some useful information.

> reduced the number of clicks required to navigate ... Yes it does

Then this issue is addressed. I’m comfortable closing it.
