### 2017-11-08 ljvmiranda921/pyswarms#61 @ljvmiranda921
Thanks a lot, @stsievert !

Merged #64
