### 2017-11-09 #433 @ljvmiranda921
**Update:** The latest commit resolves ljvmiranda921/pyswarms#61
