### 2017-11-11 #433 @kyleniemeyer
Hi @oesteban, have you had a chance to look at the software package and paper, if you are still able to?
