### 2017-11-21 453fb38
Change shape of offset to (n_particles, 2).
