### 2017-11-21 1b7297a
Fix global minimum in beale function docstring
