### 2018-01-09 #433 @kyleniemeyer
Hi @ljvmiranda921, sorry for the delay on this.

It looks like your software is good to go, but I do have some feedback on the article before accepting.

Right now, the article is a bit short; we don't want a full-length paper, but per the [author guidelines](http://joss.theoj.org/about#author_guidelines) we do expect between 250-1000 words. Perhaps you could describe the implementation a bit more, or better yet explain some example use cases. Examples of the software being used in research (whether published already, or in progress) are also helpful. In addition, it may be helpful to explain PSO in a sentence or two, with an appropriate reference, at the beginning.
