### 2018-01-10 ljvmiranda921/pyswarms#70 @ljvmiranda921
More explicit documentation in paper.md
Reference: openjournals/joss-reviews#433

> Right now, the article is a bit short; we don't want a full-length paper, but per the author guidelines we do expect between 250-1000 words. Perhaps you could describe the implementation a bit more, or better yet explain some example use cases. Examples of the software being used in research (whether published already, or in progress) are also helpful. In addition, it may be helpful to explain PSO in a sentence or two, with an appropriate reference, at the beginning.
