### 2018-01-10 47f76a7
Update paper.md (#70)
Reference:  openjournals/joss-reviews#433

Extends the previous version of the paper (very short) into
a longer and more explicit version. A short introduction for
PSO was added, some design principles, and a brief mention
to some use-cases.

Author: @ljvmiranda921
E-mail: ljvmiranda@gmail.com
