### 2018-01-10 #433 @ljvmiranda921
Hi @kyleniemeyer , thanks a lot! I have updated the paper and the changes can be seen in  ljvmiranda921/pyswarms#71 . If everything looks good, I can already merge the branch to master. :smile:
