### 2018-01-10 e1aefef
Fix multiple inline citations (#70)
