### 2018-01-10 #433 @kyleniemeyer
@ljvmiranda921 looks good! Please merge that, and I'll try generating the article PDF.
