### 2018-01-10 #433 @kyleniemeyer
@ljvmiranda921 alright, looks good. Can you now archive the entire software repository (e.g., using Zenodo) and report the DOI back here? That'll be the last thing needed.
