### 2018-01-10 #433 @ljvmiranda921
Thanks! Hmmm, I currently have a Zenodo DOI of the latest version v.0.1.7 ([link](https://zenodo.org/record/996029), DOI: 10.5281/zenodo.996029). Do I still need to make a new one?
