### 2018-01-10 #433 @kyleniemeyer
It should reflect the latest version including the changes you made to address the reviewer comments here—if that version was archived in September, then I don't think it would have the newer changes you made.
