### 2018-01-10 #433 @ljvmiranda921
Hi @kyleniemeyer , sorry for my confusion and I just want to be careful: if I am to archive via Zenodo, then that means I will create a new release, thus bumping the version number of my submission (just following the instructions from this [link](https://guides.github.com/activities/citable-code/)). This will then be v0.1.8 and is now "different" to v0.1.7, is it okay to proceed? Thanks a lot! :+1:
