### 2018-01-10 #433 @ljvmiranda921
Thank you @kyleniemeyer for helping me improve my paper, and to @stsievert for reviewing my submission! :smiley:
