### 2018-01-11 #433 @arfon
@stsievert - many thanks for your review and to @kyleniemeyer for editing this submission ✨. @ljvmiranda921 - your submission is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00433 ⚡️ 🚀 💥
